Update handlers.py
Hello, in the last few days the library PyPDF2 has been giving the following error: "PdfFileReader is deprecated and was removed in PyPDF2 3.0.0. Use PdfReader instead". So, I want to kindly recommend you change the call of PyPDF2 library to an updated one like PyPDF4. Or another solution that could work is to change the function PdfFileReader to PdfReader as the error recommend doing. Thank you and hope the changes could be done.